+++ src/main/java/org/folio/bulkops/domain/entity/BulkOperationDataProcessing.java
@@ -1,11 +1,11 @@
-package org.folio.bo.domain.entity;
+package org.folio.bulkops.domain.entity;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.With;
-import org.folio.bo.domain.dto.UpdateActionType;
+import org.folio.bulkops.domain.dto.StatusType;
 
 import javax.persistence.Entity;
 import javax.persistence.EnumType;
@@ -14,6 +14,7 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.Table;
+import java.time.LocalDateTime;
 import java.util.UUID;
 
 @Data
@@ -22,18 +23,19 @@ import java.util.UUID;
 @AllArgsConstructor
 @NoArgsConstructor
 @Entity
-@Table(name = "bulk_operation_rule_details")
-public class BulkOperationRuleDetails {
+@Table(name = "bulk_operation_data_processing")
+public class BulkOperationDataProcessing {
   @Id
   @GeneratedValue(strategy = GenerationType.AUTO)
   private UUID id;
 
-  private UUID ruleId;
+  private UUID bulkOperationId;
 
   @Enumerated(EnumType.STRING)
-  private UpdateActionType updateAction;
+  private StatusType status;
 
-  private String updateValue;
-
-
+  private LocalDateTime startTime;
+  private LocalDateTime endTime;
+  private Integer totalNumOfRecords;
+  private Integer processedNumOfRecords;
 }
