+++ src/main/java/org/folio/bulkops/domain/entity/BulkOperationExecution.java
@@ -1,11 +1,11 @@
-package org.folio.bo.domain.entity;
+package org.folio.bulkops.domain.entity;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.With;
-import org.folio.bo.domain.dto.UpdateOptionType;
+import org.folio.bulkops.domain.dto.StatusType;
 
 import javax.persistence.Entity;
 import javax.persistence.EnumType;
@@ -14,23 +14,27 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.Table;
+import java.time.LocalDateTime;
 import java.util.UUID;
 
 @Data
 @Builder
 @With
-@AllArgsConstructor
 @NoArgsConstructor
+@AllArgsConstructor
 @Entity
-@Table(name = "bulk_operation_rule")
-public class BulkOperationRule {
+@Table(name = "bulk_operation_execution")
+public class BulkOperationExecution {
   @Id
   @GeneratedValue(strategy = GenerationType.AUTO)
   private UUID id;
 
   private UUID bulkOperationId;
   private UUID userId;
+  private LocalDateTime startTime;
+  private LocalDateTime endTime;
+  private Integer processedRecords;
 
   @Enumerated(EnumType.STRING)
-  private UpdateOptionType updateOption;
+  private StatusType status;
 }
