+++ src/main/java/org/folio/bulkops/domain/entity/BulkOperationExecutionContent.java
@@ -1,11 +1,11 @@
-package org.folio.bo.domain.entity;
+package org.folio.bulkops.domain.entity;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.With;
-import org.folio.bo.domain.dto.StatusType;
+import org.folio.bulkops.domain.dto.StateType;
 
 import javax.persistence.Entity;
 import javax.persistence.EnumType;
@@ -14,7 +14,6 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.Table;
-import java.time.LocalDateTime;
 import java.util.UUID;
 
 @Data
@@ -23,19 +22,18 @@ import java.util.UUID;
 @AllArgsConstructor
 @NoArgsConstructor
 @Entity
-@Table(name = "bulk_operation_data_processing")
-public class BulkOperationDataProcessing {
+@Table(name = "bulk_operation_execution_content")
+public class BulkOperationExecutionContent {
   @Id
   @GeneratedValue(strategy = GenerationType.AUTO)
   private UUID id;
 
+  private String customIdentifier;
+  private UUID bulkOperationExecutionChunkId;
   private UUID bulkOperationId;
 
   @Enumerated(EnumType.STRING)
-  private StatusType status;
+  private StateType state;
 
-  private LocalDateTime startTime;
-  private LocalDateTime endTime;
-  private Integer totalNumOfRecords;
-  private Integer processedNumOfRecords;
+  private String errorMessage;
 }
