+++ src/main/java/org/folio/bulkops/domain/entity/BulkOperationRule.java
@@ -1,11 +1,11 @@
-package org.folio.bo.domain.entity;
+package org.folio.bulkops.domain.entity;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.With;
-import org.folio.bo.domain.dto.StateType;
+import org.folio.bulkops.domain.dto.UpdateOptionType;
 
 import javax.persistence.Entity;
 import javax.persistence.EnumType;
@@ -14,7 +14,6 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.Table;
-import java.time.LocalDateTime;
 import java.util.UUID;
 
 @Data
@@ -23,21 +22,15 @@ import java.util.UUID;
 @AllArgsConstructor
 @NoArgsConstructor
 @Entity
-@Table(name = "bulk_operation_execution_chunk")
-public class BulkOperationExecutionChunk {
+@Table(name = "bulk_operation_rule")
+public class BulkOperationRule {
   @Id
   @GeneratedValue(strategy = GenerationType.AUTO)
   private UUID id;
 
-  private UUID bulkOperationExecutionId;
   private UUID bulkOperationId;
-  private Integer firstRecordIndex;
-  private Integer lastRecordIndex;
-  private LocalDateTime startTime;
-  private LocalDateTime endTime;
+  private UUID userId;
 
   @Enumerated(EnumType.STRING)
-  private StateType state;
-
-  private String errorMessage;
+  private UpdateOptionType updateOption;
 }
