+++ src/main/java/org/folio/bulkops/repository/BulkOperationExecutionChunkRepository.java
@@ -1,6 +1,6 @@
-package org.folio.bo.repository;
+package org.folio.bulkops.repository;
 
-import org.folio.bo.domain.entity.BulkOperationExecutionChunk;
+import org.folio.bulkops.domain.entity.BulkOperationExecutionChunk;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
