+++ src/main/java/org/folio/bulkops/repository/BulkOperationExecutionRepository.java
@@ -1,6 +1,6 @@
-package org.folio.bo.repository;
+package org.folio.bulkops.repository;
 
-import org.folio.bo.domain.entity.BulkOperationExecution;
+import org.folio.bulkops.domain.entity.BulkOperationExecution;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
