+++ src/main/java/org/folio/bulkops/repository/BulkOperationRuleDetailsRepository.java
@@ -1,6 +1,6 @@
-package org.folio.bo.repository;
+package org.folio.bulkops.repository;
 
-import org.folio.bo.domain.entity.BulkOperationRuleDetails;
+import org.folio.bulkops.domain.entity.BulkOperationRuleDetails;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
